# docs/openapi/xuanchu-v1.yaml
openapi: 3.0.3
info:
  title: Xuanchu HTTP API
  version: v1
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: xuanchu_pat or xuanchu_agent
    serverAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: xuanchu_admin
  schemas:
...
    AdminSetupResult:
      type: object
      required: [token, token_name, token_prefix]
      properties:
        token:
          type: string
          description: Raw xuanchu_admin token. Returned once.
        token_name:
          type: string
        token_prefix:
          type: string
...
    Success:
      type: object
      required: [data]
      properties:
        data:
          nullable: true
...
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    BadRequest:
      description: Invalid JSON, invalid path parameter, or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Route or resource not found. Project-scoped task misses use task_not_found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource already exists or conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
paths:
...
  /api/v1/admin/setup:
    post:
      security: []
      summary: Create the first server admin token with a setup-code.
      description: Available only when `[server.admin]` is enabled and no valid admin token verifier exists. The raw admin token is returned once and its SHA-256 verifier is stored in `server_admin_tokens`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [setup_code]
              properties:
                setup_code:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: First admin token created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/AdminSetupResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/session:
    get:
      summary: Validate server admin token.
      description: Uses a server admin token from `server_admin_tokens` or the deprecated `[server.admin.tokens]` compatibility config. Returns `admin_setup_required` when admin is enabled but no valid admin token verifier exists.
      security:
        - serverAdminAuth: []
      responses:
        '200':
          description: Admin session.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/AdminSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/workspaces:
    post:
      summary: Create workspace via server admin bootstrap token.
      description: Uses a server admin token from `server_admin_tokens` or the deprecated `[server.admin.tokens]` compatibility config, not a PAT or Agent token. The admin token cannot access ordinary API endpoints.
      security:
        - serverAdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, owner]
              properties:
                slug:
                  type: string
                name:
                  type: string
                description:
                  type: string
                visibility:
                  type: string
                  enum: [private, team, public]
                owner:
                  type: object
                  required: [name]
                  properties:
                    name:
                      type: string
                    email:
                      type: string
      responses:
        '201':
          description: Workspace created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/workspaces/{workspace}/agent-tokens:
    post:
      summary: Create workspace-scoped agent token via server admin bootstrap token.
      description: The raw agent token is returned once. Server admin token hash/verifier is never returned.
      security:
        - serverAdminAuth: []
      parameters:
        - in: path
          name: workspace
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                user:
                  type: string
                  description: User ref. Defaults to the workspace owner when omitted.
                scopes:
                  type: array
                  items:
                    type: string
                project_refs:
                  type: array
                  items:
                    type: string
                project_ids:
                  type: array
                  items:
                    type: string
                expires_in:
                  type: string
                  description: Go time.ParseDuration expression, for example 24h.
                expires_in_seconds:
                  type: integer
                  format: int64
      responses:
        '201':
          description: Agent token created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/me:
    get:
      summary: Current actor, token, visible workspaces, and effective workspace.
      responses:
        '200':
          description: Actor metadata.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/tasks:
    get:
      summary: List tasks or run a report in the effective workspace.
      parameters:
        - $ref: '#/components/parameters/WorkspaceQuery'
        - $ref: '#/components/parameters/ProjectQuery'
        - $ref: '#/components/parameters/ProjectIDQuery'
        - in: query
          name: report
          schema:
            type: string
        - in: query
          name: target
          schema:
            type: string
        - in: query
          name: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Taskwarrior-style filter expression. Alias of filter. Supports assignee filters such as assignee:alice and assignee:me.
        - in: query
          name: filter
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: no_context
          schema:
            type: boolean
          description: Bypass the server-side active context for this request.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum number of tasks to return. Defaults to 200.
      responses:
        '200':
          description: Task list.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create a task.
      parameters:
        - $ref: '#/components/parameters/WorkspaceQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                project:
                  type: string
                project_id:
                  type: string
                priority:
                  type: string
                due:
                  type: integer
                  format: int64
                assignees:
                  type: array
                  items:
                    type: string
                depends:
                  type: array
                  items:
                    type: string
                wait:
                  type: integer
                  format: int64
                scheduled:
                  type: integer
                  format: int64
                until:
                  type: integer
                  format: int64
                recur:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                udas:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '201':
          description: Task created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks/{taskRef}:
    parameters:
      - in: path
        name: taskRef
        required: true
        description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
        schema:
          type: string
      - $ref: '#/components/parameters/WorkspaceQuery'
    get:
      summary: Get one task by UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
      responses:
        '200':
          description: Task details.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Modify one task by UUID or task_slug.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                project:
                  type: string
                project_id:
                  type: string
                priority:
                  type: string
                clear_project:
                  type: boolean
                clear_priority:
                  type: boolean
                due:
                  type: integer
                  format: int64
                clear_due:
                  type: boolean
                wait:
                  type: integer
                  format: int64
                clear_wait:
                  type: boolean
                scheduled:
                  type: integer
                  format: int64
                clear_scheduled:
                  type: boolean
                until:
                  type: integer
                  format: int64
                clear_until:
                  type: boolean
                assignees:
                  type: array
                  items:
                    type: string
                remove_assignees:
                  type: array
                  items:
                    type: string
                clear_assignees:
                  type: boolean
                depends:
                  type: array
                  items:
                    type: string
                clear_depends:
                  type: boolean
                recur:
                  type: string
                clear_recur:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
                remove_tags:
                  type: array
                  items:
                    type: string
                udas:
                  type: object
                  additionalProperties:
                    type: string
                clear_udas:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Modified task.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Mark a task deleted.
      responses:
        '200':
          description: Deleted task.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tasks/{taskRef}/done:
    post:
      summary: Complete a task by UUID or task_slug.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Completed task.
  /api/v1/tasks/{taskRef}/start:
    post:
      summary: Start a task by UUID or task_slug.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Started task.
  /api/v1/tasks/{taskRef}/stop:
    post:
      summary: Stop a task by UUID or task_slug.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Stopped task.
  /api/v1/tasks/{taskRef}/annotations:
    post:
      summary: Add an annotation by UUID or task_slug.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Annotated task.
  /api/v1/tasks/{taskRef}/annotations/{annotationID}:
    delete:
      summary: Remove an annotation by annotation ID.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - in: path
          name: annotationID
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Annotation removed.
  /api/v1/tasks/{taskRef}/urgency:
    get:
      summary: Explain task urgency by UUID or task_slug.
      parameters:
        - in: path
          name: taskRef
          required: true
          description: Task UUID or task_slug. Numeric working-set IDs are rejected with task_ref_invalid.
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Urgency explanation.
  /api/v1/reports/{name}:
    get:
      summary: Run a named report.
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/WorkspaceQuery'
        - $ref: '#/components/parameters/ProjectQuery'
        - $ref: '#/components/parameters/ProjectIDQuery'
        - in: query
          name: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - in: query
          name: filter
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Report task list.
  /api/v1/users:
    get:
      summary: List users.
      responses:
        '200':
          description: User list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
    post:
      summary: Create a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
  /api/v1/users/{user}:
    parameters:
      - in: path
        name: user
        required: true
        schema:
          type: string
        description: User ID or name.
    get:
      summary: Get user details.
      responses:
        '200':
          description: User details.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
  /api/v1/me/active_workspace:
    put:
      summary: Switch the actor's active workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace]
              properties:
                workspace:
                  type: string
                  description: Workspace slug or UUID.
      responses:
        '200':
          description: Active workspace switched.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
  /api/v1/workspaces:
    get:
      summary: List visible workspaces.
      parameters:
        - in: query
          name: all
          schema:
            type: boolean
      responses:
        '200':
          description: Workspace list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Workspace'
    post:
      summary: Create a workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
                name:
                  type: string
                description:
                  type: string
                visibility:
                  type: string
      responses:
        '201':
          description: Workspace created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
...
  /api/v1/workspaces/{workspace}/archive:
    post:
      summary: Archive a workspace.
      parameters:
        - in: path
          name: workspace
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace archived.
...
  /api/v1/contexts/none:
    post:
      summary: Clear active context for actor and effective workspace.
      parameters:
        - $ref: '#/components/parameters/WorkspaceQuery'
      responses:
        '200':
          description: Active context cleared.
...